+++ formidable/__init__.py
@@ -1,5 +1,5 @@
 from .json_migrations import latest_version
 
 default_app_config = 'formidable.app.FormidableConfig'
-version = '6.1.0.dev0'
+version = '6.1.0'
 json_version = latest_version