+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ master (unreleased)
 ===================
 
 - Pass context to serializer in from_json.
+- Fix license reference in setup.py (#402).
 
 Release 6.0.0 (2020-10-07)
 ==========================